removed duplicate code that was in main :(
