+++ src/main.rs
@@ -7,17 +7,10 @@ mod os_reader;
 use anyhow::Result;
 use chrono::prelude::*;
 use controller::MogRun;
-// use csv::*;
-// use cv_utils::GameWindow;
 use data_utils::PlayerHistory;
 use enigo::*;
 use os_reader::read_inputs_from_os;
-// use serde::Serialize;
-// use std::fs::OpenOptions;
-// use std::path::PathBuf;
-// use std::time::Duration;
 
-// use winput::message_loop::{self, EventReceiver};
 // const COMPASS_TIK: i32 = 381;
 // const REFRESH_RATE: u64 = 20; // game should be more like 16ms, this means we're slower
 // ingame constants if required...
@@ -96,39 +89,4 @@ fn main() -> Result<()> {
     println!("END_TIME: {:^40}", Utc::now().format("%H:%M:%S %D%m%Y"));
     println!("--------------------------------------------------------------");
     Ok(())
-}
-
-fn get_median(v: &mut Vec<usize>) -> f32 {
-    if v.len() < 1 {
-        return 0.0;
-    }
-
-    let mut vec = v.clone();
-    vec.sort();
-    if vec.len() % 2 == 1 {
-        return *vec.get(vec.len() / 2).unwrap() as f32;
-    }
-    return (*vec.get(vec.len() / 2 - 1).unwrap() + *vec.get(vec.len() / 2).unwrap()) as f32 / 2.0;
-}
-
-fn get_mode(slice: &[usize]) -> HashMap<&usize, i32> {
-    let mut map = HashMap::with_capacity(slice.len());
-    if slice.is_empty() {
-        return map;
-    }
-
-    for num in slice {
-        let count = map.entry(num).or_insert(0);
-        *count += 1;
-    }
-    let _max_value: i32 = map.values().map(|v| *v).max().unwrap();
-    map
-}
-
-fn get_mean(slice: &[usize]) -> f32 {
-    if slice.len() < 1 {
-        return 0.0;
-    }
-    let sum: usize = slice.iter().sum();
-    return sum as f32 / slice.len() as f32;
 }
