+++ src/os_reader.rs
@@ -62,28 +62,3 @@ pub fn check_monitors() {
         );
     }
 }
-
-pub fn check_elden_ring_is_running(enigo: &mut Enigo, gamemenu: &GameMenus) -> Result<bool> {
-    // Checks for something like this : 19064 eldenring.exe
-    let s = System::new_all();
-    // for (pid, process) in s.processes() {
-    //     println!("{} {}", pid, process.name());
-    // }
-    for (_, process) in s.processes() {
-        if process.name().contains("eldenring.exe") {
-            println!("Elden Ring is running");
-            return Ok(true);
-        }
-    }
-    println!("Elden Ring is not running");
-    launch_elden_ring(enigo, &gamemenu);
-    Ok(false)
-}
-
-fn launch_elden_ring(enigo: &mut Enigo, game: &GameMenus) {
-    println!("Launching eldenring.exe");
-    let _output = Command::new(r"E:\SteamLibrary\steamapps\common\ELDEN RING\Game\eldenring.exe")
-        .output()
-        .expect("failed to run eldenring.exe");
-    game.enter_game_from_main_menu(enigo)
-}
